make SSF platform independent

+++ src/main.h
@@ -713,7 +713,7 @@ int64_t get_mpow_ms_correction (CBlockIndex * p);
 bool update_ssf (int nVersion);
 
 /* Calculate the subsidy scaling factor for the CBlockIndex pointer */
-double get_ssf (CBlockIndex * pindex);
+unsigned int get_ssf (CBlockIndex * pindex);
 
 /* Get the number of blocks since the last update of the subsidy scaling factor */
 int get_ssf_height (const CBlockIndex * pindex);
